Reduce clutter in Loader by splitting up functionality in task-specific source files
